add better parameter description

+++ flask/data/api/user.py
@@ -8,21 +8,24 @@ bp = Blueprint('user', __name__, url_prefix='/api')
 
 
 @bp.route('/user', methods=['GET'])
-def get_all_user():
-    all_user = User.get_all()
-    result = users_schema.dump(all_user)
+def get_user():
 
-    return jsonify(result.data), 200
+    username = request.args.get('username', default='', type=str)
+    all = request.args.get('all', default=False, type=bool)
 
+    if all:
+        all_user = User.get_all()
+        result = users_schema.dump(all_user)
 
-@bp.route('/user/<username>', methods=['GET'])
-@permission_needed
-def get_user(username):
-    user = User.query.get(username)
-    if not user:
-        return jsonify(message='user not found'), 404
+        return jsonify(result.data), 200
 
-    return user_schema.jsonify(user), 200
+    if not all:
+
+        user = User.query.get(username)
+        if not user:
+            return jsonify(message='user not found'), 404
+
+        return user_schema.jsonify(user), 200
 
 
 @bp.route('/user/register', methods=['POST'])
@@ -39,23 +42,35 @@ def add_user():
     return make_response(jsonify(status='success')), 200
 
 
-@bp.route('/user/<username>', methods=['PUT'])
-def user_update(username):
+@bp.route('/user', methods=['PUT'])
+def user_update():
+
+    username = request.args.get('username', default='', type=str)
 
     user = User.query.get(username)
 
-    username = request.json['username']
+    if not user:
+        return make_response(jsonify(message='user not found')), 400
 
-    user.username = username
+    new_email = request.json['email']
+
+    user.email = new_email
 
     db.session.commit()
 
     return make_response(jsonify(message='user updated')), 200
 
 
-@bp.route('/user/<username>', methods=['DELETE'])
-def user_delete(username):
+@bp.route('/user', methods=['DELETE'])
+def user_delete():
+
+    username = request.args.get('username', default='', type=str)
+
     user = User.query.get(username)
+
+    if not user:
+        return make_response(jsonify(message='user not found')), 400
+
     user.delete()
 
     return make_response(jsonify(message='user deleted')), 200
